+++ lua/obsidian/command.lua
@@ -77,7 +77,7 @@ M.complete_args_search = function(client, _, cmd_line, _)
 
   local completions = {}
   local search_lwr = string.lower(search_)
-  for note in iter(client:search(search_)) do
+  for note in iter(client:search(search_, true)) do
     local note_path = assert(client:vault_relative_path(note.path))
     if string.find(note:display_name(), search_lwr, 1, true) then
       table.insert(completions, note:display_name() .. "  " .. note_path)
@@ -342,11 +342,25 @@ M.register("ObsidianBacklinks", {
     local ok, backlinks = pcall(function()
       return require("obsidian.backlinks").new(client, nil, nil, note)
     end)
+
     if ok then
-      log.info(
-        ("Showing backlinks '%s'. Hit ENTER on a line to follow the backlink."):format(tostring(backlinks.note.id))
-      )
-      backlinks:view()
+      backlinks:view(function(matches)
+        if not vim.tbl_isempty(matches) then
+          log.info(
+            "Showing backlinks to '%s'.\n\n"
+              .. "TIPS:\n\n"
+              .. "- Hit ENTER on a match to follow the backlink\n"
+              .. "- Hit ENTER on a group header to toggle the fold, or use normal fold mappings",
+            backlinks.note.id
+          )
+        else
+          if note ~= nil then
+            log.warn("No backlinks to '%s'", note.id)
+          else
+            log.warn "No backlinks to current note"
+          end
+        end
+      end)
     else
       log.err "Backlinks command can only be used from a valid note"
     end
@@ -1025,9 +1039,17 @@ M.register("ObsidianRename", {
       end, path)
     end
 
-    search.search_async(client.dir, reference_forms, { "--fixed-strings", "-m=1" }, on_search_match, function(_)
-      all_tasks_submitted = true
-    end)
+    search.search_async(
+      client.dir,
+      reference_forms,
+      nil,
+      nil,
+      { "--fixed-strings", "-m=1" },
+      on_search_match,
+      function(_)
+        all_tasks_submitted = true
+      end
+    )
 
     -- Wait for all tasks to get submitted.
     vim.wait(2000, function()